chore: use correct terminologies like digest and vc
