+++ src/4.0/exports/digest.ts
@@ -1,0 +1,1 @@
+export { digestVC, digestVCs, wrapDocumentErrors } from "../digest";
