+++ src/4.0/exports/index.ts
@@ -1,7 +1,7 @@
-export * from "./wrap";
+export * from "./digest";
 export * from "./sign";
 export * from "./obfuscate";
-export * from "./verify";
+export * from "./validate";
 export * from "./utils";
 export * from "./types";
 export * from "./builder";
