+++ src/4.0/exports/validate.ts
@@ -1,0 +1,1 @@
+export { validateDigest } from "../validate";
