--- src/4.0/exports/verify.ts
@@ -1,1 +1,0 @@
-export { verify as verifySignature } from "../verify";
