--- src/4.0/exports/wrap.ts
@@ -1,1 +1,0 @@
-export { wrapDocument, wrapDocuments, wrapDocumentErrors } from "../wrap";
